#68 useCodelist: renamed Codelist to Classification
