+++ src/components/subsetDraft/Classification.jsx
@@ -10,24 +10,37 @@ import {
     XSquare
 } from 'react-feather';
 import {Paragraph, Text, Title} from '@statisticsnorway/ssb-component-library';
-import {useGet} from '../../controllers/klass-api';
+import {useGet, useClassification} from '../../controllers/klass-api';
 import '../../css/panel.css';
 import {useTranslation} from 'react-i18next';
 import { replaceRefWithHTMLAndSanitize } from '../../utils/strings';
 
-/*
- * TODO: Use links delivered by API, do not parse - less coupling
- */
 export const Classification = ({item = {}, update, remove, from, to}) => {
-    const { t } = useTranslation();
+    const {t} = useTranslation();
 
     item.id = item._links.self.href.split('/').pop();
+    // TODO use fallback and loader
+    const {metadata, codesWithNotes} = useClassification(item.id);
 
-    const check = {
-        hasCodes: () => (item.codes && item.codes.length > 0),
-        includible: () => (!item.included && item.codes && item.codes.length > 0)
-    };
+    // TODO use fallback and loader
+    // FIXME show errors
+    const url = from && to
+        ? `/classifications/${item.id}/codes.json?from=${from}&to=${to}`
+        : `/classifications/${item.id}/codesAt.json?date=${from || to}`;
+    const [codes] = useGet(item.codes ? null : url);
+    useEffect(() => {
+        if (codes) {item.codes = codes.codes;}
+    }, [codes]);
 
+    useEffect(() => {
+        if (codes && codesWithNotes) {
+            item.codes.forEach(code =>
+                code.notes = codesWithNotes.find(c => code.name === c.name)?.notes || []
+            );
+        }
+    }, [codes, codesWithNotes]);
+
+    // TODO: outsource to useToggle()
     const toggle = {
         closeAll: () => ({
             showAlert: false,
@@ -63,121 +76,77 @@ export const Classification = ({item = {}, update, remove, from, to}) => {
 
     const [expander, setExpander] = useState(toggle.closeAll());
 
-    // TODO use fallback and loader
-    // FIXME show errors
-    const [version, isLoading, error, setVersionPath] = useGet();
-    useEffect(() => {
-        if (version && item && item.versions) {
-            let origin = item.versions.find(v => v._links.self.href = version._links.self.href);
-            if (origin) {
-                origin.classificationItems = version.classificationItems;
-                origin.classificationItems.forEach(ci => {
-                    if (ci.notes) {
-                        let code = item.codes?.find(c => c.code === ci.code);
-                        if (code) {
-                            code.notes = code.notes || [];
-                            const exists = code.notes.find(n => n.note.localeCompare(ci.notes) === 0);
-                            if (!exists) {
-                                code.notes.push({
-                                    note: ci.notes,
-                                    versionName: version.name,
-                                    validFrom: version.validFrom,
-                                    validTo: version.validTo
-                                });
-                            }
-                        }
-                    }
-                });
-            }
-        }
-    }, [version]);
-
-    // TODO use fallback and loader
-    // FIXME show errors
-    const [info] = useGet(`/classifications/${item.id}`);
-
-    useEffect(() => {
-        if (info && info.versions && info.versions.length > 0) {
-            item.versions = info.versions;
-            item.versions.forEach(v => {
-                v.id = v._links.self.href.split('/').pop();
-                setVersionPath(`/versions/${v.id}`);
-            });
-        }
-    }, [info]);
-
-    // TODO use fallback and loader
-    // FIXME show errors
-    let url = from && to
-        ? `/classifications/${item.id}/codes.json?from=${from}&to=${to}`
-        : `/classifications/${item.id}/codesAt.json?date=${from || to}`;
-    const [codes] = useGet(item.codes ? null : url);
-    useEffect(() => {
-        if (codes) {item.codes = codes.codes;}
-    }, [codes]);
+    const check = {
+        hasCodes: () => (item.codes && item.codes.length > 0),
+        includible: () => (!item.included && item.codes && item.codes.length > 0)
+    };
 
     return (
         <>
-        <div style={{display: 'flex'}}>
-            <div style={{width: '400px'}}>{item.name}</div>
+            <div style={{display: 'flex'}}>
+                <div style={{width: '400px'}}>{item.name}</div>
 
-            <button onClick={() => item.error && setExpander(toggle.alert())}>
-                <Alert color={item.error ? 'orange' : 'transparent'}/>
-            </button>
+                <button onClick={() => item.error && setExpander(toggle.alert())}>
+                    <Alert color={item.error ? 'orange' : 'transparent'}/>
+                </button>
 
-            {/*TODO: (test case) remove empty classification from draft.classification*/}
-            <button onClick={() => {
-                if (item.included || check.hasCodes()) {
-                    setExpander(toggle.closeAll());
-                    item.included = !item.included;
-                    update();
-                } else {setExpander(toggle.cannot());}}}>
-                {check.includible() && <PlusSquare color='#1A9D49'/>}
-                {item.included && <MinusSquare color='#B6E8B8'/>}
-                {!item.included && !check.hasCodes() && <XSquare color='#9272FC' />}
-            </button>
+                {/*TODO: (test case) remove empty classification from draft.classification*/}
+                <button onClick={() => {
+                    if (item.included || check.hasCodes()) {
+                        setExpander(toggle.closeAll());
+                        item.included = !item.included;
+                        update();
+                    } else {
+                        setExpander(toggle.cannot());
+                    }
+                }}>
+                    {check.includible() && <PlusSquare color='#1A9D49'/>}
+                    {item.included && <MinusSquare color='#B6E8B8'/>}
+                    {!item.included && !check.hasCodes() && <XSquare color='#9272FC'/>}
+                </button>
 
-            <button onClick={() => setExpander(toggle.codes())}>
-                <ListIcon color={check.hasCodes() ? '#3396D2' : '#C3DCDC'}/>
-            </button>
+                <button onClick={() => setExpander(toggle.codes())}>
+                    <ListIcon color={check.hasCodes() ? '#3396D2' : '#C3DCDC'}/>
+                </button>
 
-            <button onClick={() => setExpander(toggle.info())}>
-                <Info color={info ? '#62919A': '#C3DCDC'}/>
-            </button>
+                <button onClick={() => setExpander(toggle.info())}>
+                    <Info color={metadata ? '#62919A' : '#C3DCDC'}/>
+                </button>
 
-            {remove &&
+                {remove &&
                 <button onClick={() => {
                     setExpander(toggle.closeAll());
                     remove();
                 }}>
                     <Trash2 color='#ED5935'/>
                 </button>
-            }
-        </div>
+                }
+            </div>
 
-        {expander.showAlert &&
+            {expander.showAlert &&
             <div style={{backgroundColor: 'AntiqueWhite'}}
                  className='panel'><Text>{item.error}</Text>
             </div>}
 
-        {expander.showCannot &&
+            {expander.showCannot &&
             <div style={{backgroundColor: '#ece6fe'}}
                  className='panel'>
                 <Text>{t('Code list cannot be added to the subset due to lack of codes')}</Text>
             </div>}
 
-        {/* TODO limit the height and scroll*/}
-        {expander.showCodes && <Codes from={from} to={to} id={item.id}
-                                      codes={item.codes}
-                                      include={(o) => {
-                                          item.included = o ? true : item.included;
-                                          update();
-                                      }} />
+            {/* TODO limit the height and scroll*/}
+            {expander.showCodes && <Codes from={from} to={to} id={item.id}
+                                          codes={item.codes}
+                                          include={(o) => {
+                                              item.included = o ? true : item.included;
+                                              update();
+                                          }}/>
             }
 
-        {/* TODO limit the height and scroll*/}
-        {expander.showInfo && <ClassificationInfo id={item.id} info={info}/>}
-    </>);
+            {/* TODO limit the height and scroll*/}
+            {expander.showInfo && <CodelistInfo id={item.id} info={metadata}/>}
+        </>
+    );
 };
 
 export const Codes = ({from, to, codes=[], id, include}) => {
@@ -188,10 +157,10 @@ export const Codes = ({from, to, codes=[], id, include}) => {
             <div className="ssb-checkbox-group">
                 <div className="checkbox-group-header">{t('Codes')}
                     {from && to
-                    ? ` ${t('from to', { from, to })}:`
-                    : from || to ? ` ${t('at', { date: from || to})}:`
-                    : ` (${t('Period is not set').toLocaleLowerCase()})`
-                }</div>
+                        ? ` ${t('from to', { from, to })}:`
+                        : from || to ? ` ${t('at', { date: from || to})}:`
+                            : ` (${t('Period is not set').toLocaleLowerCase()})`
+                    }</div>
                 {!codes || codes.length < 1
                     ? <Text>{t('No codes found for this validity period')}</Text>
                     : <>
@@ -216,8 +185,8 @@ export const Codes = ({from, to, codes=[], id, include}) => {
                         {codes.map((code, i) =>
                             <CodeInfo key={i} id={id} item={code}
                                       onChange={() => {
-                                           code.included = !code.included;
-                                           include(code.included);
+                                          code.included = !code.included;
+                                          include(code.included);
                                       }}
                             />)
                         }
@@ -236,26 +205,26 @@ export const CodeInfo = ({id, item, onChange}) => {
     return (
         <>
             <div style={{display: 'flex'}}>
-            <div className="ssb-checkbox">
-                <input id={`${item.code}-${id}`}
-                       type='checkbox' name='include'
-                       checked={item.included}
-                       value={item.code}
-                       onChange={onChange}/>
-                <label className='checkbox-label'
-                       htmlFor={`${item.code}-${id}`}>
-                    <Text><strong>{item.code}</strong> {item.name}</Text>
-                </label>
-            </div>
-            <button onClick={() => {
-                setShowNotes(!showNotes);
-            }}><MoreHorizontal color={item.notes ? '#62919A': '#C3DCDC' }/>
-            </button>
+                <div className="ssb-checkbox">
+                    <input id={`${item.code}-${id}`}
+                           type='checkbox' name='include'
+                           checked={item.included}
+                           value={item.code}
+                           onChange={onChange}/>
+                    <label className='checkbox-label'
+                           htmlFor={`${item.code}-${id}`}>
+                        <Text><strong>{item.code}</strong> {item.name}</Text>
+                    </label>
+                </div>
+                <button onClick={() => {
+                    setShowNotes(!showNotes);
+                }}><MoreHorizontal color={item.notes ? '#62919A': '#C3DCDC' }/>
+                </button>
             </div>
 
             {showNotes && <div>
-                {!item.notes
-                    ? <Text>{t('Notes are not found.')}</Text>
+                {!item.notes || item.notes.length === 0
+                    ? <Text>{t('No notes found.')}</Text>
                     : item.notes.map(note => (
                         <div key={note} style={{
                             padding: '10px 50px 20px 50px'
@@ -264,18 +233,19 @@ export const CodeInfo = ({id, item, onChange}) => {
                             <div style={{width: '65%'}}
                                  className='ssb-paragraph'
                                  dangerouslySetInnerHTML={
-                                     {__html: replaceRefWithHTMLAndSanitize(note.note) }
+                                     {__html: replaceRefWithHTMLAndSanitize(note.note)}
                                  }
                             />
-                            <Text small><strong>«{note.versionName}»</strong> ({t('valid')}: {note.validFrom || '...'} - {note.validTo || '...'})</Text>
+                            <Text
+                                small><strong>«{note.versionName}»</strong> ({t('valid')}: {note.validFrom || '...'} - {note.validTo || '...'})</Text>
                         </div>))}
-                </div>
+            </div>
             }
         </>
     );
 };
 
-export const ClassificationInfo = ({id, info}) => {
+export const CodelistInfo = ({id, info}) => {
     const { t } = useTranslation();
 
     return (
@@ -285,18 +255,18 @@ export const ClassificationInfo = ({id, info}) => {
             <Paragraph><strong>Id:</strong> {id}</Paragraph>
             <table>
                 <tbody>
-                    <tr>
-                        <th>{t('From')}</th>
-                        <th>{t('To')}</th>
-                        <th>{t('Version')}</th>
+                <tr>
+                    <th>{t('From')}</th>
+                    <th>{t('To')}</th>
+                    <th>{t('Version')}</th>
+                </tr>
+                {info.versions.map((version, i) => (
+                    <tr key={i}>
+                        <td>{version.validFrom || '...'}</td>
+                        <td>{version.validTo || '...'}</td>
+                        <td style={{width: '65%'}}>{version.name}</td>
                     </tr>
-                    {info.versions.map((version, i) => (
-                        <tr key={i}>
-                            <td>{version.validFrom || '...'}</td>
-                            <td>{version.validTo || '...'}</td>
-                            <td style={{width: '65%'}}>{version.name}</td>
-                        </tr>
-                    ))}
+                ))}
                 </tbody>
             </table>
             <Paragraph><strong>{t('Description')}:</strong> {info.description || '-'}</Paragraph>
